elisp: gigasecond, WIP
I'm almost there, but there's a slight time difference that I can't quite figure out how to adjust. Most likely related to DST and/or time zones.

`decode-time` returns a list of numbers, the last of which is a difference in seconds from GMT. So, the solution might be to convert to seconds, then add the difference in seconds, then decode AGAIN, then drop the last 2 arguments.